% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.linwords.r
\name{am.linstems}
\alias{am.linstems}
\title{Count percentage of common stems to first string}
\usage{
am.linstems(term1, term2, minlen = 2)
}
\arguments{
\item{term1}{character vector}

\item{term2}{character vector}

\item{minlen}{number: minimal word length}
}
\value{
Number. Percentage (share) of common words between two input strings
}
\description{
Count percentage of common stems to the length of first input string. Please remove extra spaces before usage to get the right answer. Not vectorized.
}
\seealso{
am.linwords, am.linwordsu
}

