% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.linwords.r
\name{am.linwords}
\alias{am.linwords}
\title{Count percentage of common words}
\usage{
am.linwords(term1, term2)
}
\arguments{
\item{term1}{character vector}

\item{term2}{character vector}
}
\value{
Number. Percentage (share) of common words between two input strings
}
\description{
Count percentage of common words between two strings. Please remove extra spaces before usage. Vectorized.
}
\seealso{
am.linwordsu, am.linstems
}

