% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.linwords.r
\name{am.linwordsu}
\alias{am.linwordsu}
\title{Count percentage of common words to first string}
\usage{
am.linwordsu(term1, term2)
}
\arguments{
\item{term1}{character vector}

\item{term2}{character vector}
}
\value{
Number. Percentage (share) of common words between two input strings
}
\description{
Count percentage of common words to the length of first input string. Please remove extra spaces before usage to get the right answer. Vectorized.
}
\seealso{
am.linwords, am.stems
}

