% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.pairs.r
\name{am.pairs}
\alias{am.pairs}
\title{Get word pairs}
\usage{
am.pairs(term1, term2, prfx = "", minlen = 2)
}
\arguments{
\item{term1}{character vector}

\item{term2}{character vector}

\item{prfx}{character: prefix for the returned word pairs (optional)}

\item{minlen}{number: minimal word length}
}
\value{
Character
}
\description{
Get word pairs from two strings
}
\seealso{
...
}

