% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.rmwords.r
\name{am.rmwords}
\alias{am.rmwords}
\title{Remove words}
\usage{
am.rmwords(vc, words)
}
\arguments{
\item{vc}{character vector of strings to modify}

\item{words}{character vector of words to remove}
}
\value{
Character vector
}
\description{
Remove specified words from character vector
}
\seealso{
...
}

