% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.stopwords.en.r
\name{am.stopwords.en}
\alias{am.stopwords.en}
\title{English stopwords}
\usage{
am.stopwords.en()
}
\value{
Character vector of stopwords. Source: http://xpo6.com/list-of-english-stop-words/
}
\description{
Get English stopwords
}
\seealso{
...
}

